Avoid repeated work in RemoveUnusedModuleElements::addReferences() [NFC] (#7407)
Globals that refer to globals lead to more work. This work cannot be
infinite, as globals only refer to previous ones, but this can end up as
exponential time, so this is actually important to optimize here.

For exponential time, it is enough to have a chain of these:

 (global $global$N+1 (ref $A) (struct.new $A
  (global.get $global$N)
  (global.get $global$N)
 ))

That is, two references from each global to its predecessor, causing
us to double the work each time we scan back.

Fixes #7405 (where the above pattern appears)