Allow using mimalloc with dynamic linking (#7391)
With dynamic linking, build and link mimalloc's dynamic library, and  include it
in the installation (this also brings along the headers and CMake files, but it
seemed like more trouble than it was worth to try to manually install just the
library or remove the extras).
The static build remains the same.

Remove the restriction that mimalloc can only be linked into a static-lib build.