[NFC] printHeapType => printHeapTypeName in Print.cpp (#7443)
Rename the function in anticipation of exact heap types appearing in the
IR. When an expression like `StructNew` has an exact heap type, the
`exact` does not appear in `struct.new $foo`. In this case `$foo` is not
the full heap type, but rather than name of the heap type definition.