Use standard type printing for BINARYEN_PRINT_FULL (#7447)
`printTypeOrName`, used to print types for BINARYEN_PRINT_FULL and a few
other niche use cases, previously had bespoke printing for references
that did not follow the standard format. To improve the output and
reduce the number of ways we print types, change it to use the standard
printing utility.