[GC] Fix global handling in TypeRefiningGUFA (#7451)
TypeRefiningGUFA refines struct types based on the types that flow into fields,
and looks past the immediate types in the IR. As a result, it can require
casts when we end up storing something that appears not-sufficiently-
refined to wasm validation rules, in ways that cannot occur with normal
TypeRefining.

The specific problem that can happen is that casts are disallowed in
globals, so we must be careful to not refine too much there.