Revert exact HeapTypes
Revert the following PRs that introduced exactness on HeapType:

 - #7446
 - #7444
 - #7432
 - #7412
 - #7396

Keep only the changes to wasm-type-printing.cpp to fix the assertions
that subclasses of TypeNameGeneratorBase correctly override getNames.

Although putting exactness on heap types makes the most sense in the
Custom Descriptors spec, it is not a great fit for how HeapType is used
in Binaryen. In almost all cases, HeapType is used to represent heap
type definitions rather than the dynamic types of values. Letting
HeapType represent exact heap types is not useful in those cases and
opens up a new class of possible bugs.

To avoid these bugs, we had been introducing a new HeapTypeDef type to
represent heap type definitions, but nearly all uses of HeapType would
have had to have been replaced with HeapTypeDef. Rather than introduce
HeapTypeDef as a third type alongside Type and HeapType, it will be
simpler to continue using HeapType to represent heap type definitions
and Type to represent the dynamic types of values, including their
exactness. While exactness will syntactically be part of the heap type,
it will be part of the `Type` in Binaryen IR.

A follow-on PR will restore the original work on exact references, and
PRs following that one will update the encoding, parsing, and printing
of exact references to match the current spec.