[NFC] Add an explicit ModuleRunner.start() (#7463)
Previously the constructor ran the start method and other initialization. Doing
that in an explicit function that the user calls makes it possible to do things in
the middle. The specific thing I would like to do in the middle is have a function
to change the interpreter instance's mode to "reject as nonconstant relaxed SIMD",
but I've wanted this in the past too, and worked around it - seems best to just
fix it.

Adding more flags to the constructor is another option, but there are already
several, and more in the parent classes, and several levels of inheritance through
which all such options must be forwarded, which is annoying.