Switch source map parsing to use the JSON parser (#7467)
This is in preparation for fixing #6805.
Aside from cleaning up and reusing more code, it allows optional fields
in the
source map and random access to those fields, rather than walking
through
the whole JSON file just once. (It does keep the current behavior of
walking through the actual mappings in a single pass synchronized with
reading
the binary). It also adds some unit tests.