[NFC] Represent exactness and sharedness with one bit (#7500)
Since the Type representation only needs to include a bit for sharedness
for basic heap types and only needs to include a bit for exactness for
non-basic heap types, they can use the same bit. Using fewer bits in the
Type representation is beneficial because it reduces the minimum
alignment of HeapTypeInfos required to avoid trampling the reserved
bits.