[SIMD] wasm-ctor-eval: Fix relaxed SIMD in expressions (#7508)
The previous fix #7495 handled functions computing something with
relaxed SIMD, in which case an exception is thrown, but we also need to
handle relaxed SIMD inside a function, since wasm-ctor-eval can
partially-evaluate code. In that case we get NONCONSTANT_FLOW.

Turns out adding --kept-exports is enough to test this, as keeping the
exports around forces partial evaluation of their contents.