[Fuzzing] Simplify V8 fuzzing flags: use --fuzzing over specific EH flag (#7509)
ClusterFuzz already applies this flag, and it implies the EH flag, so we might
as well do it the way V8 does it. This is simpler, though it does mean we are
testing something even more different than production - but in the way V8
believes is best for fuzzing, at least.