Disallow exact casts without custom descriptors enabled
When we emit modules containing exact types without custom descriptors
enabled, we generalize those types to be inexact so the written module
is valid for the feature set. This fudging of the types changes the
behavior of casts, though; casts to exact types that should have failed
at runtime may in fact succeed after the binary is written and the cast
target type is fudged to be inexact. To avoid the binary writer
introducing this change in behavior, disallow casts to exact types when
custom descriptors is not enabled.

This will require optimizations to avoid introducing casts to exact
types when custom descriptors is not enabled. We hypothesize that this
will be simpler than the alternative, which would be to have all
optimizations (and the interpreter) treat casts to exact types as casts
to inexact types when custom descriptors is disabled.

Test the new validation errors and and update an existing test that this
change invalidates by splitting off the part that depended on custom
descriptors being disabled into a separate test.