Always allow trivial exact casts (#7516)
We usually disallow casts to exact types when custom descriptors are
disallowed to prevent the cast semantics changing when the binary writer
makes the cast types inexact. At the same time, we plan to allow exact
types to appear elsewhere in the IR even when custom descriptors are not
enabled. This means that the input to a cast may have an exact type.
Finalization of casts sets the cast target to be the GLB of the old cast
target and the input type, so this could produce invalid exact casts.

However, this situation only arises when the cast will trivially succeed
because the cast target is a supertype of the input type. If the cast
input were exact and the cast target were not a supertype of the input
type, then the GLB would be an inexact bottom type, so there would be no
problem.

Since these trivial casts would remain trivial after binary writing
removes exactness, it is safe to allow them to validate even when custom
descriptors is disabled. Make this change to avoid finalization
producing invalid casts. The alternative would be to add logic to cast
finalization to avoid introducing exactness, but that would both be more
complicated and would lose type information, harming optimization power.