Allow null-checking exact casts without custom descriptors
We previously allowed trivial exact casts (i.e. those where the input
and target types are the same) to validate even without custom
descriptors enabled to ensure that finalization always produces valid
casts. But validation can also produce casts where the cast target is a
non-nullable exact reference and the input type is a nullable exact
reference to the same heap type. Update validation to allow these casts.

This is safe because erasing the exactness of the input and cast types
does not change the results of these casts. They succeed iff the input
is a non-null value.