Fuzzer: Avoid constantly growing the wasm each time (#7511)
Imagine we start with a fuzz file, then mutate it (using it as initial
content and letting the fuzzer make changes, perhaps with
`--preserve-imports-and-exports`). We can mutate it again and
again, which is what harnesses like Fuzzilli do. It is somewhat bad
if we keep growing the file, because then when we explore the
space of wasm files, the further we travel, the more we focus on
big files for no good reason.

Running `--fuzz-passes` is one way to trim the wasm (it might end
up running `-O3`), but not reliable. This PR goes through places
that always add code and tries to at least give a chance to not do
so. As long as there is a chance to not grow, then a harness can
see two testcases of equal coverage and pick the smaller one.

Specific changes:

* If a data segment exists already, do not always add others.
* If an exnref table exists already, reuse it (like we do with funcref).
* Don't always add hashMemory support (which is large).
* Avoid errors in the fuzzer on wasm files without exports. That is
  now possible, since we no longer always add at least 1 export.