[NFC] Simplify Array2Struct type replacement (#7539)
Array2Struct has to update the types of every expression that interacts
with and produces a reference to the optimized array type. It previously
did this by separately checking whether a nullable or non-nullable
reference to the array was a subtype of the expression's type. Simplify
this logic by doing only a single check that considers only the heap
types of the references.

Also remove some unnecessary variables in which various reference types
were cached since it is extremely cheap to materialize a reference type
now.

These simplifications will also make it easier to update the pass to
handle exact reference types once `array.new` instructions are typed as
exact.