[NFC] Simplify interpretation of casts
The interpreter previously inspected subtyping for nullability and heap
types separately when executing casts. Simplify this to just check
subtyping on the combined type. This is NFC for now, but will have the
happy side effect of handling exactness correctly once we type
allocations as exact and interpret exact casts.