Avoid errors in Type::with(HeapType)
Previously doing e.g. `type.with(HeapType::none)` would cause an
assertion failure if `type` was exact because `.with()` would only
replace the heap type and exact references to basic heap types are
disallowed.

Rather than checking for and avoiding this error in all the callers,
simply drop exactness when `.with()` is called with a basic heap type.
This is reasonable behavior because the only alternative is never
correct.

Add a test that hits an assertion failure without this fix.
AbstractTypeRefining replaces a defined type with `none` and the type
updating utility does not check whether the new heap type is basic
before doing the replacement.