Handle exactness in MinimizeRecGroups (#7555)
Treat rec groups differing only in the exactness of a reference as
different, but only when custom descriptors is enabled. When custom
descriptors is not enabled, exactness will be erased before the binary
is written, so if two minimized rec groups differed only in exactness,
they would in fact be written as the same rec group. This would change
the behavior of casts meant to differentiate between types in that rec
group, so it would be incorrect.