[NFC] Use a SmallVector for HeapType children (#7565)
This was one of danleh's top findings in the mimalloc investigation,
where this method accounted for 25% (!) of all allocations.

I measured various sizes of SmallVectors and indeed it is possible to
get noticeably faster here: this PR is a 5% speedup for -O3 as a whole,
tested on a large Java testcase and a large Kotlin testcase.

The change to MinimizeRecGroups is NFC (the order is reversed, but it
doesn't matters). This is needed to compile, as our current SmallVector
doesn't have reverse iteration support (I looked into that for a few minutes
but it was not trivial to add; anyhow, the new code is idiomatic in the
codebase, I think).