[Code Annotations] Start Branch Hinting support, in the text format, for Breaks (#7567)
This adds the first minimal amount of work for Branch Hinting (and
code annotations, more generally): Just the text format, and just a
single instruction (Break).

This design follows debug info: Assuming most instructions have no
branch hints (like most do not have debug info), we do not add a field
to the Break class itself, but to metadata on the side.

---------

Co-authored-by: Thomas Lively <tlively123@gmail.com>