[NFC] Cleanly separate expression tracking from source maps (#7571)
We track debug info in two ways: for source maps, and for DWARF. DWARF needs
more information, not just the start of each expression's location but also the end
and delimiters (else for an if) as well. The existing code slightly mixed the two
together, which is annoying because code annotations require the same tracking
DWARF does. To improve that, this PR refactors the code to cleanly separate the
two forms of tracking:

* writeSourceMapLocation is now the only method that source maps use.
* trackExpressionStart|End|Delimiter is now used by DWARF (and soon code
  annotations).

As a result,

* BinaryInstWriter no longer needs a sourceMap param. It was using !sourceMap
  in the sense of "maybe DWARF", but given custom annotations we'll need more
  anyhow. Simplify the code by letting the parent decide what to do.
  Replace DWARF scanning ahead of sections with a "preScan" method. This will
  be extended for code annotations later.