[NFC] Move optimal-LEB emitting logic to a shared place (#7575)
This allows generating sections in side buffers, rather than just on the
main output buffer ("o"). This will help code annotations, where it is
useful to emit them to a side buffer, then copy them into the right
place (which is before the code section, but we must emit the code
first, to know the offsets for the annotations, so reordering is
unavoidable).