[NFC] Properties::hasUnwritableTypeImmediate
Many WasmGC instructions have type index immediates that determine the
static types of certain operands. When we emit the IR, we need to look
at these operands to determine the type index immediate to emit. In
cases where the operand is unreachable or has bottom type, there is no
possible type index to emit, so we have to emit a replacement.

Previously Print.cpp had a visitor for each instruction that has such a
type immediate to individually check whether the replacement was
necessary. Refactor this to use a new
`Properties::hasUnwritableTypeImmediate` function implemented on top of
a new DELEGATE_FIELD_REF_CHILD macro in wasm-delegations-fields.def.
This function will be used elsewhere in a following PR.