Update TypeMerging for exact types
TypeMerging was already careful not to merge a subtype and its supertype
where there is a cast to the subtype that could differentiate it from
its supertype. Now with exact types, it's possible to have an exact cast
to the supertype that can differentiate it from the subtype as well.
Update the pass to collect types used as the target of exact casts and
ensure they are not merged with their subtypes.

It would be natural to use a single map from HeapType to bool to track
the cast target types and whether each one is used in an exact cast, but
that would regress performance because this pass previously used a
SmallSet to track cast types. I spent a few hours trying to create a
SmallMap that shared most of its code with SmallSet, but it would
require more time to make that work. For now, just use a second SmallSet
to track exact exact casts.