Update TypeSSA for exact types (#7579)
TypeSSA replaces allocations with allocations of new subtypes of the
original allocated type. This is not valid when the exactness of the
original allocation can be observed, e.g. because the allocation can
flow into some other instruction that would not be valid unless the
allocation is exactly the original type.

Fix TypeSSA for the presence of exact types by using ChildTyper to find
instructions that require a child to be an exact reference. When such a
child is found, inhibit all optimization of allocations for that child's
type. This is the best we can do without performing a data flow analysis
to prove that individual allocations do not flow into locations where
their exactness can be observed.