Fix BranchHinting + source maps (#7585)
We move the Code section around when writing BranchHints (since we emit
the Code first to find the offsets, then put BranchHints before it as the spec
requires). Source map offsets are not relative to the Code section but
absolute to the start of the binary, so we must adjust.