Fix tests of refinalization after optimization of selects
According to their comments, these tests purported to show that refinalize was
run after optimizing a select to a single arm with a more refined type. However,
the tests were flawed because the select arms had the same type, meaning the
parsed select node did not have a less refined type than the arms. Fix this by
giving the arms different types. Use a block to directly show that
refinalization occurred instead of the cast that was there previously.