Add PublicTypeValidator for use in fuzzer (#7597)
When custom descriptors is not enabled, we consider it a validation
error for a public type to include an exact reference because the
identity of that type will change when the binary is written and that
reference becomes inexact. Once the fuzzer observes functions signatures
with exact references, it will have to be careful not to export those
functions and make the exact reference public when custom descriptors is
disabled.

Add a utility to efficiently determine whether a type is valid to be
made public, memoizing results to avoid traversing any type definition
more than once. Use this utility in the fuzzer.