Fix assertion failure in type-refining-gufa
`filterPackedDataReads` had an assertion that the read field existed,
but it was possible for this to not be the case when the reference being
read from was null. Avoid the assertion failure by checking for this
case and filtering the read results to `none` because a read from null
will never return.