[Strings] Fix StringLowering of reference-taken functions (#7613)
We manually update the types of functions that have size 1 rec groups,
because we need to update imports and exports, even though that
changes the ABI - something TypeMapper does not do, but StringLowering
wants. See the existing comment for details.

That code updated those types, but did just themselves - it did not go
through the code to find RefFuncs of them, which left them with the old
type, causing errors. To fix this, tell TypeUpdater to map those types (it
doesn't know it needs to automatically, since we manually handled them).