Fix IRBuilder location tracking of If starts (#7617)
The old code forgot the If scope when it started the Else, so any if-else would get
a binary location span that started in the else (and ends in the right place).

To fix this, note the if start location when we have it, and don't trample.

Fixes branch hints on if-elses.