Export std hash template specializations from dylibs (#7618)
Use explicit visibility to force std::hash template specializations to
be
exported from libbinaryen dynamic library. 

Currently
we are just using the default flags, causing most of our symbols to have
"default" visibility, meaning they can all be used from the tool
sources.
However a recent libc++ change caused functions declared in namespace
std
(e.g. hash template specializations) to have hidden visibility,
inherited.
from the namespece (see
https://github.com/llvm/llvm-project/pull/131156).
So this macro forces them to be exported. Currently it is only applied
to
the hash specializations that are defined in libbinaryen and used in the
tool code. In the future if we want to compile libbinaryen with
-fvisibility-hidden or use a DLL on Windows, we'll need
to explicitly annotate everything we want to export. But that's probably
only useful if we want external users to link against libbinaryen.so
(currently we don't; it's only used to reduce our install size).