[custom-descriptors] Branching descriptor casts
Implement basic support for `br_on_cast_desc` and `br_on_cast_desc_fail`
as new variants of `BrOn`. Include binary and text parsing and printing
as well as validation. Also handle the new operations anywhere the
compiler would otherwise complain about a non-exhaustive switch over the
`BrCastOp` enum.

For validation of type immediates during parsing, relax the requirement
that the cast type is a subtype of the input type. Continue validating
in the IR that the non-descriptor branching casts are downcasts, but do
not validate this for the new descriptor casts. See
https://github.com/WebAssembly/custom-descriptors/issues/37 for context.

Add new validation and parsing tests in the form of spec tests in
preparation for creating an upstream spec test suite. While we're at it,
move the ref.get_desc tests to spec tests, add a few test cases, and fix
a bug when the ref.get_desc operand is null.