Fix assertion failure in TupleOptimization
We previously counted a local use for all `local.tee` operands to
`tuple.extract`, even if the `local.tee` was unreachable and had a
non-tuple local. This led to an assertion failure later. Fix the issue
by disregarding `local.tee` operands of `tuple.extract` if they don't
have tuple locals.