Fix wabt setup script (#7655)
On Linux we use "ubuntu" as the platform string when searching for a
wabt release tarball and search for a file that ends with
"-ubuntu.tar.gz".

However wabt release tarballs mention Ubuntu version as well, e.g. the
current release name is "wabt-1.0.37-ubuntu-20.04.tar.gz". The current
check with `endswith` does not match the current release tarball names.

Update the name matching to use the regex `^wabt-.*-ubuntu-.*.tar.gz$`,
which matches version numbers after the "ubuntu" part.