[Custom Descriptors] Interpret ref.get_desc (#7675)
Add a descriptor to the GCData stored in a reference value. The
descriptor itself is another literal representing either a null if there
is no descriptor or otherwise a reference to the descriptor. Update
the interpretation of struct.new to set the descriptor when allocating
types with descriptors and then retrieve the descriptor when
interpreting ref.get_desc.