[Custom Descriptors] Interpret descriptor casts
Descriptor casts must check that the expected descriptor is not null,
check whether the cast value is null, and if so, whether the cast allows
null values through, and finally must check whether the descriptor on
a non-null cast value matches the expected descriptor. This is fairly
different than the implementation of a normal cast, but the possible
results are the same. Add a new `doDescCast` helper mirroring the
existing `doCast` helper and use it to implement the various descriptor
casts.