Fix UB in OptimizeInstructions (#7676)
When optimizing `add` operations, OptimizeInstructions calculates the
max bits of the operands and shifts that value to produce a bit mask.
When an operand is unreachable, it is arbitrarily considered to have a
max bits value of 64, even for 32-bit operations. This leads to UB in
the shift. Fix the problem by returning before the shift in this case.