[NFC] Refactor heap type fuzzer generator (#7681)
Refactor the top-level driver in the heap type generator to plan
recursion groups before planning the types they contain. This will make
it easier to add support for generating described and descriptor types
in a follow-on PR because these types can only be added if there is
sufficient room left in the current recursion group.