[NFC] Generalize DFS machinery in makeInhabitable
The heap type fuzzer provides a utility for transforming a type
definition graph to ensure that all the defined types are inhabitable.
This requires finding and breaking cycles of non-nullable references. We
use a simple DFS to find the cycles, but this DFS previously assumed
that each type would be visited only once at a time because it would
always make the first backedge found nullable.

A follow-on PR will update this DFS to handle descriptor types, which
are like non-nullable references that cannot be made nullable. When the
DFS finds a backedge correspnding to a descriptor, it will not be able
to make that edge nullable. Instead, it will have to continue searching
for the next backedge. To make that follow-on PR simpler, refactor the
data structures used in the DFS to make it easier to adapt to visiting a
type more than once at a time.