[Custom Descriptors] Descriptors in type fuzzer (#7686)
Generate custom descriptor types in the heap type fuzzer. When
generating an eligible new type, choose a length for its descriptor
chain. The length is limited by the space remaining in the current
recursion group. Save the described type in a list of other such
described types, and optionally pick a type to describe from that list
when generating future types. Take descriptor chains into account to
ensure validity when choosing supertypes as well.

Also update other parts of the type fuzzer to acccount for custom
descriptors. For example, update the utility for making types
inhabitable to consider custom descriptors as another kind of
non-nullable reference.

Disable custom descriptors in the main fuzzer because it cannot yet
handle generated custom descriptor types.