[Custom Descriptors] Keep trapping initializers
Update RemoveUnusedModuleElements to preserve traps in global and
element segment initializers due to null or possibly-null descriptors
passed to struct.new. For every struct.new that appears with a
descriptor in an initializer, find the origin of the descrpitor value.
If it is a null literal or a nullable import, keep the global containing
the struct.new because it either will or will possibly trap.