[Branch Hinting] Add branch hint handling in RemoveUnusedBrs (#7706)
Add some utilities for easily updating/copying/clearing branch hints, then use
those in the pass.

As a drive-by, move flip() from wasm-builder.h, which everyone was including
but only this one cpp file was using, and update it in the new location (this
avoids including the new branch hints header in a central place).