[Branch Hinting] Fix Inlining's flipping of ifs, and add hint copying (#7714)
Split inlining will flip conditions in the copied code.

To do this, add support for copying code annotations when we
copy instructions (this is the first PR that can test that). This is done
in a new metadata.h helper, that subsumes part of the previous
debuginfo.h, that is, the new metadata::copyBetweenFunctions
will copy all metadata, both debug info and code annotations.