[Custom Descriptors] Optimize ref.cast_desc (#7718)
Update the cast optimizations in OptimizeInstructions to correctly
optimize ref.cast_desc. The optimizations differ from the optimizations
of normal casts because they have to be careful not to skip any side
effects in the descriptor operand and because type information alone is
not sufficient to prove that a cast will succeed.