Revert part of #7715: Ignore metadata in function comparisons (#7737)
Our comparisons are only used in optimization atm, and we decided not to
consider metadata when merging functions, as that is what LLVM does: it
will fold code together without carefully checking if profiling info matches,
and without even trying to fix up that profiling info later.

For more on LLVM, see discussion in #7733, but basically LLVM will just merge
two instructions with different branch_weights and keep the first of the two,
no matter what.