Improve ReFinalize::replaceUntaken
ReFinalize replaces unreachable branches with unreachable blocks so that
it can possibly propagate the unreachability up to the former branch
target. It previously only did this for branches whose sent values (as
opposed to their conditions) were unreachable, except for BrOn, where it
could erroneously optimize when the sent value was reachable and the
descriptor was unreachable. This was incorrect because `replaceUntaken`
assumed that the sent value was unreachable.

Rather than fix the bug for BrOn, generalize `replaceUntaken` to work
no matter whether the sent value or the other child exists or is the
source of unreachability.