Track indirect call types in RemoveUnusedModuleElements (#7728)
An indirect call to a type in a table now only forces functions of
that type to be marked as used: functions of other types are
left alone, potentially leaving them unreached.

This is more precise than assuming any indirect call can
reach anywhere, which is more or less what we did before.

The downside is that this makes the pass is 10% slower. This is one
of our faster passes, however, so the cost seems worth it, given
that this can noticeably help in some cases, e.g., 6.5% code size
reduction on Poppler, and 20% on an embind testcase. In
general, however, most real-world codebases benefit little if at
all, because there is too much overlap in indirect call signatures:
Statistically, when you generate random graphs of size `n` and
chance for each edge to exist `p`, then even if `p` decreases to 0 the
graph will tend to end up fully connected [1]. And, in wasm, `p`
does not even decrease to 0:

* Consider some common signature like `{i32} -> {}` (i32 param, no
result).
* In real-world code, there is some chance `q>0` for that signature to
be called,
  and some chance `r>0` for that signature to exist in the code.
* `p >= O(rq) > 0` because all it takes for a connection to exist is
that that signature exists on one side and is called on the other.

That is, in large codebases there is an overlap in signatures, and
statistically this means that all the code will end up reachable, at
least in the limit. But some programs do get lucky and benefit
from this PR.

[1]
https://en.wikipedia.org/wiki/Erd%C5%91s%E2%80%93R%C3%A9nyi_model#Properties_of_G(n,_p)