Fix skipping of nested/internal passes (#7741)
When --skip-pass is used, it identifies the pass to skip by name. We did not
update the name of nested/internal passes, so their name was empty,
causing skipping to fail (such passes are wrapped by another struct, which
had an empty name; to fix it, copy the name to the wrapper).

As a drive-by, rename the existing skip-pass tests, which mentioned -O1
even though they use -O2, and the opt level isn't even worth mentioning.