[Custom Descriptors] Fix desc casts in AbstractTypeRefining
When a described type is never allocated, AbstractTypeRefining will
optimize it to the bottom heap type `none`. However, if that type is
used as the targtet in any descriptor casts, subsequent refinalization
will restore the original type as derived from the type of the
descriptor operand. To avoid this reintroduction of optimized types,
fix up affected descriptor casts to be non-descriptor casts before
refinalizing. This is valid because we know that only null values can
ever pass these casts, so the descriptors are never used.

As a drive by, use a more efficient (and less verbose) method of
collecting the public types in AbstractTypeRefining as well.