[Metadce] Report removed imports due to RemoveUnusedModuleElements (#7748)
wasm-metadce does a graph analysis to find unreached things, and then
cleans up using RemoveUnusedModuleElements. That pass become more
powerful in #7728, which led to a situation where an import was removed
from the wasm, but wasm-metadce did not report that it had removed it.
This led to unneeded code in the JS (it kept sending that import,
unnecessarily). This was a harmless minor waste of JS size, but it did cause
a test error on Emscripten (#7747), as it parses that JS to check some
things, and it found an import in JS without a use in wasm.

To fix that, check if that pass removed imports, and report them.