[Custom Descriptors] Ensure descriptor validity in Unsubtyping (#7743)
We already partially handled the interaction between descriptors and
subtyping by ensuring that B <: A implies B.desc <: A.desc if B.desc and
A.desc both exist. However, there was a subtler situation we did not
handle:

A -> A.desc
^      ^
|    B.desc
|      ^
C -> C.desc

This subtyping violates the constraint that the descriptor of C's
immediate supertype must be the immediate supertype of C's descriptor.
To fix this, find the described type B to insert between C and A before
establishing C.desc <: B.desc <: A.desc.