[Custom Descriptors] Handle nullable ref.cast_desc in Heap2Local (#7744)
Heap2Local previously assumed that if an optimized allocation flowed
into a ref.cast_desc as the descriptor operand, the cast must fail
because the same allocation is known not to be the the descriptor of the
cast value. This did not account for the case where the cast value is
null and the cast allows nulls, though. Fix the output in that case to
cast the value to null and avoid trapping where the original descriptor
cast would not have trapped.