[BranchHints] Fuzz branch hints (#7704)
Add two helper passes, one to delete specific branch hints by their
instrumentation ID (as added by InstrumentBranchHints), and one to
remove all instrumentation. The new fuzzer then

* Adds random branch hints
* Instruments them and runs that to see the output
* Delete all incorrect hints
* Remove all instrumentation, leaving a wasm with correct hints only
* Optimize
* Add new instrumentation and run that to see the output

The idea is that once we have a wasm with only correct hints, the
optimizer is allowed to remove some (e.g. in DCE), but it should
never emit an invalid branch hint (e.g. by forgetting to flip a hint
when it flips an if).

We do need to avoid passes that reorder or unconditionalize
code, or parts of them, so this is not quite that simple, but it
still allows us to fuzz this.